Updated undoAction and redoAction operations
